topic_repository_module/topic-repo-requirements.txt
gevent
cassandra-driver
dataclasses